+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.1 - 2023.01.28
+### Readme and more
+++ vanilla js examples
+++ better phpDoc
+++ change $_POST to Craft::$app->request->post() < better
+
 ## 2.0.0 - 2022.08.26
 ### PHP 8.0 | Craft 4